Rename modbus register header file
